Add support for Dotty (#48)
* Fix most incompatibilities

* Remove implicit and update build

* Change default scalaVersion and change override without parens so they match to avoid Dotty crash

* Fix auto-application sites and then ambiguous reference sites

* Fix additional auto-application sites and use implicit conversions at numeric sites to remove warnings

* Fix access errors and required parens

* Fix remaining qualified import issues

* Segregate Scala 2 and Dotty enum source

* Fix enum variance type problems

* Fix most other auto-application sites

* Small improvement but gives warning in Dotty

* Prepare build for dotty options and fix implicit

* Comment out line that won't compile with dotty in Java

* Fixes to allow tests to compile in Dotty

* Fix multiversal equality

* Comment out Dotty so changes can be merged

* Scalafmt

+++ sconfig/shared/src/main/scala/org/ekrich/config/impl/ConfigImplUtil.scala
@@ -56,7 +56,7 @@ object ConfigImplUtil {
         case '\t' =>
           sb.append("\\t")
         case _ =>
-          if (isC0Control(c)) sb.append("\\u%04x".format(c.toInt))
+          if (isC0Control(c.toInt)) sb.append("\\u%04x".format(c.toInt))
           else sb.append(c)
       }
       i += 1
